requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
idna==3.6
Markdown==3.5.2
pillow==10.2.0
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
